Run code at build time to determine Perl/JIT type mappings
I give up. Turns out this module ALSO needs to know the Perl/JIT
type-mappings at build time. So there's a custom "configure" stage
whether we like it or not.

This builds bootstrap/determine_jit_types, then runs it to generate a
header file that has mappings for the abstract JIT type (jit_type_NV and
friends) as well as the actual C type to use (jit_NV). On top of that,
to make the Perl-type-constant generation code simpler and more
efficient (though efficiency hardly matters there), we also define the
name of the jit_constant_t union member to use.

*sigh*

Now this also needs to be exposed to Perl::JIT. I'm still having a
raging internal argument about which is the least worst solution for
that:

- ExtUtils::Depends to ship the header (sledgehammer)
- Ship header manually using File::ShareDir (finicky)
- Also generate simple Perl code that can reproduce the header so that
  Perl::JIT can use that at configure time (hacky)

All of the above make this module a *configure*-time dependency of
Perl::JIT instead of a *build*-time dependency. But I suppose that's not
a real problem any more these days.