jit_label_undefined needs to return a new object each time
jit_label_undefined is a numeric constant, Perl sees a reference-based
object. This means that the whole IN_OUTLIST logic has been somewhat
flawed: The label is modified in-place by the label-related libjit
functions.

To be honest, I am not sure whether the semantics of this change are
exactly what we want, but at least it's a lot more practical than what
it was before - and very close to the C-level libjit API.